refactor: split Backups's Metadata marshalling logic into two classes (ES, OS)
